+++ src/ol_openedx_canvas_integration/ol_openedx_canvas_integration/settings/lms/common.py
@@ -5,3 +5,4 @@ def plugin_settings(settings):
     """Settings for the canvas integration plugin."""  # noqa: D401
     settings.CANVAS_ACCESS_TOKEN = None
     settings.CANVAS_BASE_URL = None
+    settings.CANVAS_COURSE_ID_CACHE_TIMEOUT = 300